Implement MVC - Export logic to Controllers
