+++ routes/admin.js
@@ -1,22 +1,13 @@
 const express = require("express");
 
+const productsController = require("../controllers/products");
+
 const router = express.Router();
 
-const products = [];
-
 // /admin/add-product => GET
-router.get("/add-product", (req, res) => {
-  res.render("add-product", {
-    pageTitle: "Add Product",
-    path: "/admin/add-product",
-  });
-});
+router.get("/add-product", productsController.getAddProduct);
 
 // /admin/add-product => POST
-router.post("/add-product", (req, res) => {
-  products.push({ title: req.body.title });
-  res.redirect("/");
-});
+router.post("/add-product", productsController.postAddProduct);
 
-exports.router = router;
-exports.products = products;
+module.exports = router;
